// ClientApp13/ClientAppNew/src/app/shared/models/user-login.ts
export class UserLogin {
    bankNamea?: string;
    bankNumber?: string;
    bankType:  number = 0;
    email?: string;
    firstName?: string;
    fullName?: string;
    id?: number;
    isActive?: boolean;
    lastName?: string;
    money: number = 0;
    phone?: string;
    refreshToken?: string;
    refreshTokenExpiryTime?: string;
    token?: string; 
    typeName?:  number;
    userName?: string;
    constructor(){}
}